Big commit - failing InputState
In this state I'm failing because I've decided to start taking user
input.  I want to make sure the inputstate is passed into the update
handler. This is a bit nontraditional - normally input stores a bunch of
monostate objects.  I figure since it's immutable it should be passed
in.  I can see a scenario where that's an enormous amount of overhead to
do so though, and since it's immutable the global data is less of an
issue.